Data.Type.Functor.Product

Classes

class (PFunctor f, SFunctor f, PFoldable f, SFoldable f) => FProd (f :: Type -> Type) where

Minimal complete definition

singProd, withIndices, htraverse, ixProd, toRec, withPureProd

Associated Types

type Elem f = (i :: f k -> k -> Type) | i -> f

type Prod f = (p :: (k -> Type) -> f k -> Type) | p -> f

Methods

singProd :: Sing as -> Prod f Sing as

withIndices :: Prod f g as -> Prod f (Elem f as :*: g) as

traverseProd :: forall g h as m. Applicative m => (forall a. g a -> m (h a)) -> Prod f g as -> m (Prod f h as)

zipWithProd :: (forall a. g a -> h a -> j a) -> Prod f g as -> Prod f h as -> Prod f j as

htraverse :: Applicative m => Sing ff -> (forall a. g a -> m (h (ff @@ a))) -> Prod f g as -> m (Prod f h (Fmap ff as))

ixProd :: Elem f as a -> Lens' (Prod f g as) (g a)

toRec :: Prod f g as -> Rec g (ToList as)

withPureProd :: Prod f g as -> (PureProd f as => r) -> r

type Shape f = (Prod f Proxy :: f k -> Type)

class PureProd (f :: Type -> Type) (as :: f k) where

Methods

pureProd :: (forall a. g a) -> Prod f g as

pureShape :: PureProd f as => Shape f as

class PureProdC (f :: Type -> Type) c (as :: f k) where

Methods

pureProdC :: (forall a. c a => g a) -> Prod f g as

class ReifyConstraintProd (f :: Type -> Type) c (g :: k -> Type) (as :: f k) where

Methods

reifyConstraintProd :: Prod f g as -> Prod f (Dict c :. g) as

class FProd f => ProdSing f where

Methods

prodSing :: Prod f Sing as -> Sing as

type AllConstrainedProd c as = AllConstrained c (ToList as)

Functions

indexProd :: FProd f => Elem f as a -> Prod f g as -> g a

indexSing

Arguments

:: FProd f

=> Elem f as a
Witness

-> Sing as
Collection

-> Sing a

Extract the item from the container witnessed by the Elem

singShape :: FProd f => Sing as -> Shape f as

mapProd :: FProd f => (forall a. g a -> h a) -> Prod f g as -> Prod f h as

foldMapProd :: (FProd f, Monoid m) => (forall a. g a -> m) -> Prod f g as -> m

hmap :: FProd f => Sing ff -> (forall a. g a -> h (ff @@ a)) -> Prod f g as -> Prod f h (Fmap ff as)

zipProd :: FProd f => Prod f g as -> Prod f h as -> Prod f (g :*: h) as

imapProd :: FProd f => (forall a. Elem f as a -> g a -> h a) -> Prod f g as -> Prod f h as

itraverseProd :: (FProd f, Applicative m) => (forall a. Elem f as a -> g a -> m (h a)) -> Prod f g as -> m (Prod f h as)

ifoldMapProd :: (FProd f, Monoid m) => (forall a. Elem f as a -> g a -> m) -> Prod f g as -> m

ifoldMapSing :: forall f k (as :: f k) m. (FProd f, Monoid m) => (forall a. Elem f as a -> Sing a -> m) -> Sing as -> m

foldMapUni but with access to the index.

foldMapSing :: forall f k (as :: f k) m. (FProd f, Monoid m) => (forall (a :: k). Sing a -> m) -> Sing as -> m

A foldMap over all items in a collection.

eqProd :: (FProd f, ReifyConstraintProd f Eq g as) => Prod f g as -> Prod f g as -> Bool

An implementation of equality testing for all FProd instances, as long as each of the items are instances of Eq.

compareProd :: (FProd f, ReifyConstraintProd f Ord g as) => Prod f g as -> Prod f g as -> Ordering

An implementation of order comparison for all FProd instances, as long as each of the items are instances of Ord.

Instances

data Rec (a :: u -> Type) (b :: [u]) :: forall u. (u -> Type) -> [u] -> Type where

A record is parameterized by a universe u, an interpretation f and a list of rows rs. The labels or indices of the record are given by inhabitants of the kind u; the type of values at any label r :: u is given by its interpretation f r :: *.

Constructors

RNil :: forall u (a :: u -> Type) (b :: [u]). Rec a ([] :: [u])

(:&) :: forall u (a :: u -> Type) (b :: [u]) (r :: u) (rs :: [u]). !(a r) -> !(Rec a rs) -> Rec a (r ': rs) infixr 7

data Index :: [k] -> k -> Type where

Witness an item in a type-level list by providing its index.

Constructors

IZ :: Index (a ': as) a

IS :: Index bs a -> Index (b ': bs) a

data PMaybe :: (k -> Type) -> Maybe k -> Type where

Constructors

PNothing :: PMaybe f Nothing

PJust :: f a -> PMaybe f (Just a)

data IJust :: Maybe k -> k -> Type where

Witness an item in a type-level Maybe by proving the Maybe is Just.

Constructors

IJust :: IJust (Just a) a

data PEither :: (k -> Type) -> Either j k -> Type where

Constructors

PLeft :: PEither f (Left e)

PRight :: f a -> PEither f (Right a)

data IRight :: Either j k -> k -> Type where

Witness an item in a type-level Either j by proving the Either is Right.

Constructors

IRight :: IRight (Right a) a

data NERec :: (k -> Type) -> NonEmpty k -> Type where

Constructors

(:&|) :: f a -> Rec f as -> NERec f (a :| as) infixr 5

data NEIndex :: NonEmpty k -> k -> Type where

Witness an item in a type-level NonEmpty by either indicating that it is the "head", or by providing an index in the "tail".

Constructors

NEHead :: NEIndex (a :| as) a

NETail :: Index as a -> NEIndex (b :| as) a

data PTup :: (k -> Type) -> (j, k) -> Type where

Constructors

PSnd :: f a -> PTup f '(w, a)

data ISnd :: (j, k) -> k -> Type where

Trivially witness an item in the second field of a type-level tuple.

Constructors

ISnd :: ISnd '(a, b) b

data PIdentity :: (k -> Type) -> Identity k -> Type where

Constructors

PIdentity :: f a -> PIdentity f (Identity a)

data IIdentity :: Identity k -> k -> Type where

Trivially witness the item held in an Identity.

Since: 0.1.3.0

Constructors

IId :: IIdentity (Identity x) x

sameIndexVal :: Index as a -> Index as b -> Maybe (a :~: b)

Test if two indices point to the same item in a list.

We have to return a Maybe here instead of a Decision, because it might be the case that the same item might be duplicated in a list. Therefore, even if two indices are different, we cannot prove that the values they point to are different.

sameNEIndexVal :: NEIndex as a -> NEIndex as b -> Maybe (a :~: b)

Test if two indices point to the same item in a non-empty list.

We have to return a Maybe here instead of a Decision, because it might be the case that the same item might be duplicated in a list. Therefore, even if two indices are different, we cannot prove that the values they point to are different.

Singletons

data SIndex as a :: Index as a -> Type where

Kind-indexed singleton for Index. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SIndex', which has an actual proper Sing instance.

Constructors

SIZ :: SIndex (a ': as) a IZ

SIS :: SIndex bs a i -> SIndex (b ': bs) a (IS i)

data SIJust as a :: IJust as a -> Type where

Kind-indexed singleton for IJust. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SIJust', which has an actual proper Sing instance.

This distinction will be unnecessary once Sing is a type family.

Constructors

SIJust :: SIJust (Just a) a IJust

data SIRight as a :: IRight as a -> Type where

Kind-indexed singleton for IRight. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SIRight', which has an actual proper Sing instance.

Constructors

SIRight :: SIRight (Right a) a IRight

data SNEIndex as a :: NEIndex as a -> Type where

Kind-indexed singleton for NEIndex. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SNEIndex', which has an actual proper Sing instance.

Constructors

SNEHead :: SNEIndex (a :| as) a NEHead

SNETail :: SIndex as a i -> SNEIndex (b :| as) a (NETail i)

data SISnd as a :: ISnd as a -> Type where

Kind-indexed singleton for ISnd. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SISnd', which has an actual proper Sing instance.

Constructors

SISnd :: SISnd '(a, b) b ISnd

data SIIdentity as a :: IIdentity as a -> Type where

Kind-indexed singleton for IIdentity. Provided as a separate data declaration to allow you to use these at the type level. However, the main interface is still provided through the newtype wrapper SIIdentity', which has an actual proper Sing instance.

Since: 0.1.5.0

Constructors

SIId :: SIIdentity (Identity a) a IId

data family Sing (a :: k) :: Type

The singleton kind-indexed data family.

Defunctionalization symbols

data ElemSym0 (f :: Type -> Type) :: f k ~> (k ~> Type)

data ElemSym1 (f :: Type -> Type) :: f k -> k ~> Type

type ElemSym2 (f :: Type -> Type) (as :: f k) (a :: k) = Elem f as a

data ProdSym0 (f :: Type -> Type) :: (k -> Type) ~> (f k ~> Type)

data ProdSym1 (f :: Type -> Type) :: (k -> Type) -> f k ~> Type

type ProdSym2 (f :: Type -> Type) (g :: k -> Type) (as :: f k) = Prod f g as